+++ sprockets/mixins/mediatype/transcoders.py
@@ -6,6 +6,7 @@ Content transcoding classes & functions.
 - :class:`TextContentHandler` transcoder that translates binary bodies
   to text before calling functions that encode & decode text
 - :class:`JSONTranscoder` intelligent transcoding for JSON bodies
+- :class:`MsgPackTranscoder` intelligent transcoding for MsgPack bodies
 
 The :class:`sprockets.mixins.mediatype.ContentMixin` uses transcoder
 instances *under the hood*.  A transcoder is responsible for transforming
@@ -14,10 +15,18 @@ This module implements the transcoders.
 
 """
 import base64
+import collections
 import json
 import uuid
 
 from tornado import escape
+try:
+    import umsgpack
+except ImportError as exc:  # pragma no cover
+    umsgpack = None
+
+    def MsgPackTranscoder(*args, **kwargs):
+        raise exc
 
 
 class BinaryContentHandler(object):
@@ -147,7 +156,7 @@ class JSONTranscoder(TextContentHandler):
 
     +-------------------------------+----------------------------------------+
     | Type                          | Format                                 |
-    +-------------------------------+----------------------------------------+
+    +===============================+========================================+
     | :class:`bytes`,               | Base64 encoded string.                 |
     | :class:`bytearray`,           |                                        |
     | :class:`memoryview`           |                                        |
@@ -211,3 +220,79 @@ class JSONTranscoder(TextContentHandler):
 
         """
         return json.loads(data, **self.load_options)
+
+
+if umsgpack is not None:
+
+    class MsgPackTranscoder(BinaryContentHandler):
+        """
+        MsgPack transcoder that understands basic types.
+
+        The :meth:`.normalize_datum` method is called to normalize
+        each packed datum into something that ``umsgpack`` understands.
+
+        - sequences & sets are converted to :class:`list` instances
+        - :class:`datetime.datetime` instances are converted to
+          :class:`str` instances containing the ISO8601 representation
+        - :class:`uuid.UUID` instances are converted to :class:`str`
+          instances via ``str(value)``
+        - types mentioned in :data:`PACKABLE_TYPES` are left alone
+
+        .. note::
+
+           This transcoder is only available if you have the `umsgpack`_
+           library installed.  You can install the library yourself or
+           use ``sprockets.mixins.mediatype['msgpack']`` as a dependency
+
+        .. _umsgpack: https://pypi.python.org/pypi/u-msgpack-python
+
+        """
+
+        PACKABLE_TYPES = (bool, int, float, str, bytes)
+        """Types that umsgpack understands."""
+
+        def __init__(self, content_type='application/msgpack'):
+            super(MsgPackTranscoder, self).__init__(content_type, self.packb,
+                                                    self.unpackb)
+
+        def packb(self, data):
+            """Pack `data` into a :class:`bytes` instance."""
+            return umsgpack.packb(self.normalize_datum(data))
+
+        def unpackb(self, data):
+            """Unpack :class:`object` from a :class:`bytes` instance."""
+            return umsgpack.unpackb(data)
+
+        def normalize_datum(self, datum):
+            """
+            Convert `datum` into something packable.
+
+            :param object datum: value to prepare for msgpack'ing
+            :return: a packable version of `datum`
+            :raise TypeError: if `datum` cannot be packed
+
+            """
+
+            if datum is None or isinstance(datum, self.PACKABLE_TYPES):
+                return datum
+
+            if hasattr(datum, 'strftime'):
+                return datum.strftime('%Y-%m-%dT%H:%M:%S.%f%z')
+
+            if isinstance(datum, uuid.UUID):
+                return str(datum)
+
+            if isinstance(datum, (bytearray, memoryview)):
+                return bytes(datum)
+
+            if isinstance(datum, (collections.Sequence, collections.Set)):
+                return [self.normalize_datum(item) for item in datum]
+
+            if isinstance(datum, collections.Mapping):
+                out = {}
+                for k, v in datum.items():
+                    out[k] = self.normalize_datum(v)
+                return out
+
+            raise TypeError(
+                '{} is not msgpackable'.format(datum.__class__.__name__))