+++ docs/history.rst
@@ -5,6 +5,7 @@ Version History
 ---------------
 - Add :func:`sprockets.mixins.mediatype.add_transcoder`
 - Add :class:`sprockets.mixins.mediatype.transcoders.JSONTranscoder`
+- Add :class:`sprockets.mixins.mediatype.transcoders.MsgPackTranscoder`
 
 `1.0.4`_ (14 Sep 2015)
 ----------------------